+++ src/core/lattice/SerialPopulator.h
@@ -15,15 +15,17 @@
 class SerialPopulator : public LatticePopulator {
 private:
     std::array<std::size_t, 3> axisOrder{};
+    std::size_t startFrom{};
+    std::size_t every{};
 
 public:
     /**
      * @brief Constructs the object.
      * @param axisOrderString order in which axes will be looped through.
+     * @param startFrom skip that many shapes and start filling from the next.
+     * @param every every how many lattice slots to fill (for 1, all subsequent slots are filled)
      */
-    explicit SerialPopulator(const std::string &axisOrderString)
-            : axisOrder{LatticeTraits::parseAxisOrder(axisOrderString)}
-    { }
+    explicit SerialPopulator(const std::string &axisOrderString, std::size_t startFrom = 0, std::size_t every = 1);
 
     /**
      * @brief Returns molecules for a given @a lattice populating it cell by cell and looping through them in the axis
@@ -33,7 +35,8 @@ public:
      * lattice (Y is the outermost loop). All cells but last are filled fully (the last one may not be filled fully in
      * order to meet @a numOfShapes target). Order of shapes within the cell is preserved.
      */
-    [[nodiscard]] std::vector<Shape> populateLattice(const Lattice &lattice, std::size_t numOfShapes) const override;
+    [[nodiscard]] std::vector<Shape> populateLattice(const Lattice &lattice,
+                                                     std::optional<std::size_t> numOfShapes) const override;
 };