+++ src/frontend/matchers/LatticeMatcher.cpp
@@ -24,7 +24,6 @@
 #include "core/lattice/LayerRotationTransformer.h"
 #include "core/lattice/LayerWiseCellOptimizationTransformer.h"
 #include "core/lattice/RotationRandomizingTransformer.h"
-#include "core/lattice/RotationRandomizingTransformer.h"
 
 using namespace pyon::matcher;
 
@@ -50,7 +49,7 @@ namespace {
 
     struct PopulatorData {
         std::shared_ptr<LatticePopulator> populator;
-        std::size_t numShapes{};
+        std::optional<std::size_t> numShapes;
     };
 
     class LatticePackingFactory : public PackingFactory {
@@ -58,7 +57,7 @@ namespace {
         Lattice lattice;
         std::vector<std::shared_ptr<LatticeTransformer>> transformers;
         std::shared_ptr<LatticePopulator> populator;
-        std::size_t numShapes{};
+        std::optional<std::size_t> numShapes;
 
     public:
         LatticePackingFactory(Lattice lattice, std::vector<std::shared_ptr<LatticeTransformer>> transformers,
@@ -81,6 +80,9 @@ namespace {
 
     class AutoAxisOrderSerialPopulator : public LatticePopulator {
     private:
+        std::size_t startFrom{};
+        std::size_t every{};
+
         [[nodiscard]] std::string optimizeAxisOrder(const Lattice &lattice) const {
             std::string axisOrder = "xyz";
             auto dim = lattice.getDimensions();
@@ -91,18 +93,21 @@ namespace {
         }
 
     public:
+        AutoAxisOrderSerialPopulator(std::size_t startFrom, std::size_t every) : startFrom{startFrom}, every{every}
+        { }
+
         [[nodiscard]] std::vector<Shape> populateLattice(const Lattice &lattice,
-                                                         std::size_t numOfShapes) const override
+                                                         std::optional<std::size_t> numOfShapes) const override
         {
-            SerialPopulator serialPopulator(this->optimizeAxisOrder(lattice));
+            SerialPopulator serialPopulator(this->optimizeAxisOrder(lattice), this->startFrom, this->every);
             return serialPopulator.populateLattice(lattice, numOfShapes);
         }
     };
 
     class FullPopulator : public LatticePopulator {
     public:
-        [[nodiscard]] std::vector<Shape> populateLattice(const Lattice &lattice,
-                                                         [[maybe_unused]] std::size_t numOfShapes) const override
+        [[nodiscard]] std::vector<Shape>
+        populateLattice(const Lattice &lattice, [[maybe_unused]] std::optional<std::size_t> numOfShapes) const override
         {
             return lattice.generateMolecules();
         }
@@ -222,7 +227,7 @@ namespace {
                 Lattice theLattice(cell, latticeDim);
 
                 auto transformations = do_create_transformations(kwargs);
-                PopulatorData fullPopulator{std::make_shared<FullPopulator>(), 0};
+                PopulatorData fullPopulator{std::make_shared<FullPopulator>(), std::nullopt};
                 return std::make_shared<LatticePackingFactory>(theLattice, transformations, fullPopulator);
             });
     }
@@ -376,17 +381,23 @@ namespace {
 
     MatcherDataclass create_serial() {
         auto axisOrderAuto = MatcherString("auto");
+        auto nShapes = MatcherInt{}.positive().mapTo<std::optional<std::size_t>>();
+        auto nShapesNone = MatcherNone{}.mapTo<std::optional<std::size_t>>();
 
         return MatcherDataclass("serial")
-            .arguments({{"n_shapes", MatcherInt{}.positive().mapTo<std::size_t>()},
-                        {"axis_order", axisOrderAuto | axisOrder, R"("auto")"}})
+            .arguments({{"n_shapes", nShapes | nShapesNone, "None"},
+                        {"axis_order", axisOrderAuto | axisOrder, R"("auto")"},
+                        {"start_from", MatcherInt{}.nonNegative().mapTo<std::size_t>(), "0"},
+                        {"every", MatcherInt{}.positive().mapTo<std::size_t>(), "1"}})
             .mapTo([](const DataclassData &serial) -> PopulatorData {
-                auto nShapes = serial["n_shapes"].as<std::size_t>();
+                auto nShapes = serial["n_shapes"].as<std::optional<std::size_t>>();
                 auto axisOrder = serial["axis_order"].as<std::string>();
+                auto startFrom = serial["start_from"].as<std::size_t>();
+                auto every = serial["every"].as<std::size_t>();
                 if (axisOrder == "auto")
-                    return {std::make_shared<AutoAxisOrderSerialPopulator>(), nShapes};
+                    return {std::make_shared<AutoAxisOrderSerialPopulator>(startFrom, every), nShapes};
                 else
-                    return {std::make_shared<SerialPopulator>(axisOrder), nShapes};
+                    return {std::make_shared<SerialPopulator>(axisOrder, startFrom, every), nShapes};
             });
     }
 
@@ -507,7 +518,7 @@ namespace {
         if (kwargs.hasKey("fill_partially"))
             return kwargs["fill_partially"].as<PopulatorData>();
         else
-            return {std::make_shared<FullPopulator>(), 0};
+            return {std::make_shared<FullPopulator>(), std::nullopt};
     }
 }